Initial commit
Two tables containing the sudoku problems and solutions in a relational database. Log for database problems. Database preparer still in procedural programming form, but remaining classes as in OOP form. Given sudokus can be initially checked to ensure that they do not contain two or more of the same number in a row/column/box.  The basic methods "Last Possible Number" and "Last Remaining Cell in a Box/Row/Column" are implemented and allow the three easier sudokus to be solved.